+++ flutter_reach_five/lib/helpers/password_policy_converter.dart
@@ -1,22 +1,20 @@
-import 'package:reach_five_identity_repo/reach_five_identity_repo.dart'
-    as identity_repo;
+import 'package:reach_five_identity_repo/reach_five_identity_repo.dart';
 
 import '../models/password_policy.dart';
 
-/// Used to be the link between the [identity_repo.PasswordPolicy] from the
+/// Used as the link between the [PasswordPolicyResponse] from the
 /// identity repo and the [PasswordPolicy] exported from this package
 class PasswordPolicyConverter {
-  /// convert a [identity_repo.PasswordPolicy] to a [PasswordPolicy]
-  static PasswordPolicy fromInterface(
-    identity_repo.PasswordPolicy passwordPolicy,
-  ) => PasswordPolicy(
-    minLength: passwordPolicy.minLength,
-    minStrength: passwordPolicy.minStrength,
-    specialCharacters: passwordPolicy.specialCharacters,
-    uppercaseCharacters: passwordPolicy.uppercaseCharacters,
-    lowercaseCharacters: passwordPolicy.lowercaseCharacters,
-    digitCharacters: passwordPolicy.digitCharacters,
-    allowUpdateWithAccessTokenOnly:
-        passwordPolicy.allowUpdateWithAccessTokenOnly,
-  );
+  /// convert a [PasswordPolicyResponse] to a [PasswordPolicy]
+  static PasswordPolicy fromInterface(PasswordPolicyResponse passwordPolicy) =>
+      PasswordPolicy(
+        minLength: passwordPolicy.minLength,
+        minStrength: passwordPolicy.minStrength,
+        specialCharacters: passwordPolicy.specialCharacters,
+        uppercaseCharacters: passwordPolicy.uppercaseCharacters,
+        lowercaseCharacters: passwordPolicy.lowercaseCharacters,
+        digitCharacters: passwordPolicy.digitCharacters,
+        allowUpdateWithAccessTokenOnly:
+            passwordPolicy.allowUpdateWithAccessTokenOnly,
+      );
 }