+++ reach_five_identity_repo/lib/reach_five_identity_repo.dart
@@ -12,7 +12,7 @@ export 'package:reach_five_identity_repo/src/api/email_api.dart';
 export 'package:reach_five_identity_repo/src/api/o_auth_api.dart';
 
 export 'package:reach_five_identity_repo/src/model/client_config_response.dart';
-export 'package:reach_five_identity_repo/src/model/password_policy.dart';
+export 'package:reach_five_identity_repo/src/model/password_policy_response.dart';
 export 'package:reach_five_identity_repo/src/model/revoke_token_request.dart';
 export 'package:reach_five_identity_repo/src/model/send_email_verification200_response.dart';
 export 'package:reach_five_identity_repo/src/model/send_email_verification_request.dart';